Data Science Projects/PawMatchAI/Model_Deployment/requirements.txt
numpy>=1.20.0,<2.0.0
torch>=2.0.0,<2.1.0
torchvision>=0.15.0,<0.16.0
matplotlib>=3.5.0,<4.0.0
gradio>=5.0.0,<6.0.0
diffusers>=0.23.0,<0.24.0  
transformers>=4.30.0,<4.31.0  
accelerate>=0.20.0,<0.21.0
huggingface_hub>=0.25.1,<0.30.0  
xformers>=0.0.20,<0.1.0
timm>=0.9.0,<1.0.0
Pillow>=9.0.0,<11.0.0
ultralytics

